requirements.txt
nltk
Pillow
pydantic
PyYAML
fastapi
qdrant_client
Requests
streamlit
transformers
opencv-python
tqdm
uvicorn

